+++ lib/src/cook/process/toolchain/Configuration.hpp
@@ -74,6 +74,7 @@ namespace cook { namespace process { namespace toolchain {
 
         bool add_callback(const Configuration & config);
         bool has_config(const std::string & key, const std::string & value) const;
+        bool has_config(const std::string & key) const;
         std::list<std::string> config_values(const std::string & key) const;
 
         template <typename Functor>